feat: refactoring the equipment service
