+++ Backend/Controllers/ItemController.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace Backend.Controllers
+{
+    [ApiController]
+    [Route("api/equipment")]
+    public class ItemController
+    {
+
+    }
+}
