+++ Backend/Services/EquipmentService.cs
@@ -1,71 +1,34 @@
+﻿using Core.Models.DTOs;
 
-using Backend.Interfaces;
-using Backend.Models.Entities;
-using Backend.Models.Requests;
-
-namespace Backend.Services;
-
-public class EquipmentService : IEquipmentService
+namespace Backend.Services
 {
-
-    private readonly IEquipmentRepository _equipmentRepository;
-    private readonly ILogger<EquipmentService> _logger;
-    private readonly IConfiguration _configuration;
-    public EquipmentService(IEquipmentRepository repository, ILogger<EquipmentService> logger, IConfiguration configuration)
+    public class EquipmentService : Core.Interfaces.IEquipmentService
     {
-        _equipmentRepository = repository;
-        _logger = logger;
-        _configuration = configuration;
-    }
-    
-    public async Task<List<Equipment>> GetAllEquipment()
-    {
-        return await _equipmentRepository.GetAllEquipmentAsync();
-    }
+        private readonly ILogger<EquipmentService> _logger;
 
-    public async Task<Equipment?> GetEquipment(NameRequest nameRequest)
-    {
-        // Since all names are stored as upper case, need to query via upper casing
-        string name = nameRequest.Name.ToUpperInvariant();
-
-        return await _equipmentRepository.GetEquipmentByNameAsync(name);
-    }
-
-    public async Task<Equipment?> GetEquipment(IdRequest idRequest)
-    {
-        string id = idRequest.Id;
-        return await _equipmentRepository.GetEquipmentByIdAsync(id);
-    }
-
-    public async Task<Equipment> CreateEquipment(Equipment equipment)
-    {
-        // Need to capitalize the name so its unique on all 
-        equipment.Name = equipment.Name.ToUpperInvariant();
-
-        Equipment created = await _equipmentRepository.CreateEquipmentAsync(equipment);
-        return created;
-    }
-
-    public async Task<Equipment?> UpdateEquipment(NameRequest nameRequest, Equipment equipment)
-    {
-        string name = nameRequest.Name.ToUpperInvariant();
-
-        // Check if it exists already; if not, throw an argument exception
-        Equipment? existingEquipment = await GetEquipment(nameRequest);
-
-        if (existingEquipment is null)
+        public Task<EquipmentDTO> CreateEquipment(EquipmentDTO equipmentToCreate)
         {
-            return null;
+            throw new NotImplementedException();
         }
 
-        // Otherwise, bind with the id
-        equipment.Id = existingEquipment.Id;
-        return await _equipmentRepository.UpdateEquipmentByNameAsync(name, equipment);
-    }
+        public Task<List<EquipmentDTO>> GetAllEquipment()
+        {
+            throw new NotImplementedException();
+        }
 
-    public async Task<Equipment?> UpdateEquipment(IdRequest idRequest, Equipment equipment)
-    {
-        string id = idRequest.Id;
-        return await _equipmentRepository.UpdateEquipmentByIdAsync(id, equipment);
+        public Task<EquipmentDTO> GetEquipmentByIdAsync(string id)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<EquipmentDTO> GetEquipmentByNameAsync(string name)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Task<EquipmentDTO> UpdateEquipment(EquipmentDTO equipmentToUpdateTo)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
